==> Master/1stGrade/1stSemester/DAA/Competition/process/2.1_Feature_Eng.ipynb <==
# IMPORTS

import re
import holidays
from collections import Counter
import sklearn as skl
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
from sklearn import preprocessing
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.svm import SVC
from sklearn.metrics import classification_report,plot_confusion_matrix, accuracy_score
from lightgbm import LGBMClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.neighbors import KNeighborsClassifier

# leitura
incidentes = pd.read_csv('../input/training_data.csv')
teste = pd.read_csv('../input/test_data.csv')

# remoção de colunas desnecessárias
incidentes=incidentes.drop(['city_name'],axis=1)
incidentes=incidentes.drop(['avg_precipitation'],axis=1)
teste=teste.drop(['city_name'],axis=1)
teste=teste.drop(['avg_precipitation'],axis=1)

# transformação para numéricas
incidentes["avg_rain"] = incidentes['avg_rain'].replace({'Sem Chuva' : 0, 'chuva fraca' : 1, 'chuva moderada' : 2, 'chuva forte' : 3}).astype(int)
incidentes["magnitude_of_delay"] = incidentes['magnitude_of_delay'].replace({'UNDEFINED' : 0, 'MODERATE' : 1, 'MAJOR' : 2}).astype(int)
incidentes['luminosity'] = incidentes['luminosity'].replace(['DARK','LOW_LIGHT','LIGHT'],[0,1,2]).astype(int)
incidentes['record_date'] = pd.to_datetime(incidentes['record_date'])
incidentes["incidents"] = incidentes['incidents'].replace({'None' : 0, 'Low' : 1, 'Medium' : 2, 'High' : 3, 'Very_High' : 4}).astype(int)

teste["avg_rain"] = teste['avg_rain'].replace({'Sem Chuva' : 0, 'chuva fraca' : 1, 'chuva moderada' : 2, 'chuva forte' : 3}).astype(int)
teste["magnitude_of_delay"] = teste['magnitude_of_delay'].replace({'UNDEFINED' : 0, 'MODERATE' : 1, 'MAJOR' : 2}).astype(int)
teste['luminosity'] = teste['luminosity'].replace(['DARK','LOW_LIGHT','LIGHT'],[0,1,2]).astype(int)

## Cenário 1.2 - Feature Engeneering com método diferente

#### Método diferente para tratar as estradas afetadas

# MÉTODO 2
# create the two new columns
incidentes["National"] = 0
incidentes["Regional"] = 0

def treat_affected_roads_types(df):
    # Check if the value in the "affected_roads" column is null, empty, or equal to ","
    mask = df["affected_roads"].isnull() | df["affected_roads"].eq("") | df["affected_roads"].eq(",")
    df.loc[mask, "National"] = 0
    df.loc[mask, "Regional"] = 0
    
    # For the remaining rows, split the string in the "affected_roads" column by comma and count the occurances of each road type
    for index, row in df[~mask].iterrows():
        # Create a Counter object to store the counts of each road type
        road_counts = Counter({"N": 0, "R": 0})
        
        # Split the string and update the counts in the Counter object
        roads = row["affected_roads"].split(",")
        for road in roads:
            # Use a regular expression to extract the road type
            road_type = re.search("^[NR]\d+", road)
            if road_type:
                road_type = road_type.group()[0]
                road_counts[road_type] += 1
        
        # Set the values in the new columns
        df.loc[index, "National"] = road_counts["N"]
        df.loc[index, "Regional"] = road_counts["R"]
        
    return df

incidentes = treat_affected_roads_types(incidentes)
teste = treat_affected_roads_types(teste)

# remover a coluna tratada
incidentes=incidentes.drop(['affected_roads'],axis=1)
teste=teste.drop(['affected_roads'],axis=1)

# tratar o record_time
# treino
incidentes['record_date'] = pd.to_datetime(incidentes['record_date'])
incidentes['day_of_month'] = incidentes['record_date'].dt.day
incidentes['week_day'] = incidentes['record_date'].dt.weekday
incidentes['month'] = incidentes['record_date'].dt.month
incidentes['year'] = incidentes['record_date'].dt.year
incidentes['hour'] = incidentes['record_date'].dt.hour
incidentes['minute'] = incidentes['record_date'].dt.minute
incidentes['day_of_year'] = incidentes['record_date'].dt.dayofyear

# teste
teste['record_date'] = pd.to_datetime(teste['record_date'])
teste['day_of_month'] = teste['record_date'].dt.day
teste['week_day'] = teste['record_date'].dt.weekday
teste['month'] = teste['record_date'].dt.month
teste['year'] = teste['record_date'].dt.year
teste['hour'] = teste['record_date'].dt.hour
teste['minute'] = teste['record_date'].dt.minute
teste['day_of_year'] = teste['record_date'].dt.dayofyear


# remover essa coluna
incidentes.drop('minute', axis=1, inplace=True)
incidentes.drop('year', axis=1, inplace=True)
teste.drop('minute', axis=1, inplace=True)
teste.drop('year', axis=1, inplace=True)


* Vamos extrair algumas features que podem ter interesse:
    * Estação do Ano;
    * Fim de semana;
    * Feriado;
* A ideia é depois relacionar com as métricas e verificar se o modelo melhorou

# obter estação do ano para cada dia
def get_season(df):
    seasons = []
    for index, row in df.iterrows():
        # Access the data in each column by column name
        month = row['month']
        day = row['day_of_month']
    
        # Determine the season based on the month and day of the month
        if month in [3, 4, 5]:
            # For March, April, and May, use the day to determine the season
            if day < 21:
                season = 0
            else:
                season = 1
        elif month in [6, 7, 8]:
            season = 1
        elif month in [9, 10, 11]:
            # For September, October, and November, use the day to determine the season
            if day < 21:
                season = 2
            else:
                season = 3
        else:
            season = 3
        seasons.append(season)
    return seasons

incidentes['season'] = get_season(incidentes)
teste['season'] = get_season(teste)

# determina se é ou não um dia de férias
def is_holidays(df):
    pt_holidays = holidays.CountryHoliday("PT")
    df["is_holiday"] = 0

    for i, row in df.iterrows():
        date_field = row["record_date"]
        if date_field in pt_holidays:
            df.at[i, "is_holiday"] = 1
    return df

incidentes = is_holidays(incidentes)
teste = is_holidays(teste)

# determina se é fim de semana ou não
def is_weekend(df):
    df["is_weekend"] = 0
    for i, row in df.iterrows():
        date_field = row["week_day"]
        if date_field in [0, 4]:
            df["is_weekend"] = 1
    return df

incidentes = is_weekend(incidentes)
teste = is_weekend(teste)

# podemos remover a coluna
incidentes.drop('record_date', axis=1, inplace=True)
# podemos remover a coluna
teste.drop('record_date', axis=1, inplace=True)

incidentes.head()

### Modelo
* Vamos verificar se há alterações nas métricas

#divisão do target
x = incidentes.drop(['incidents'], axis=1)
y = incidentes['incidents']

# função que constrói os modelos, apresentando as métricas
def train_and_predict(X, y, teste):
    # Split the data into training and testing sets
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

    models = [RandomForestClassifier(n_estimators=100, max_features="auto", random_state=42),
             LGBMClassifier(boosting_type='gbdt', learning_rate=0.1, n_estimators=100, random_state=42),
             GradientBoostingClassifier(learning_rate=0.1, n_estimators=100, random_state=42),
             DecisionTreeClassifier(criterion='gini', splitter='best', max_depth=20, random_state=42),
             ExtraTreesClassifier(criterion='gini', max_depth=20, random_state=42),
             KNeighborsClassifier(n_neighbors=5, weights='uniform', algorithm='auto')]
             
    reports=[]
    acc = []
    pred = []
    names = ['RF','LGB','GB','DT','ET','KN']
    
    for i, model in enumerate(models):
        # Train the model
        model.fit(X_train, y_train)
        
        # Make predictions on the test set
        y_pred = model.predict(X_test)
        report = classification_report(y_test, y_pred)
        reports.append(report)
        accuracy = accuracy_score(y_test, y_pred)
        acc.append(accuracy)
        test_predictions = model.predict(teste)
        pred.append(test_predictions)

    i = 0   
    #Print the reports side by side
    for i, report in enumerate(reports):
        print("Model {} {}:\n{}".format(i + 1, names[i], report))
        print("Accuracy: {:.5f}".format(acc[i]))
        print("=" * 50)
    return pred

preds = train_and_predict(x,y,teste)

### Comparação

| **Modelo**                          | **Antes** | **Depois** | **Obs**      | **Depois (M2)**|
|-------------------------------------|-----------|------------|--------------|----------------|
| **K Neighbors Classifier**          | 0.8312    | 0.83120    | **Manteve**  |       0.83360  |
| **Extra Trees Classifier**          | 0.8968    | 0.89760    | **Melhorou** |         0.90640|
| **Gradient Boosting Classifier**    | 0.8984    | 0.90000    | **Melhorou** |      0.90960   |
| **Decision Tree Classifier**        | 0.9080    | 0.90640    | Piorou       |         0.92240|
| **Random Forest Classifier**        | 0.9168    | 0.92480    | **Melhorou** |      0.92960   |
| **Light Gradient Boosting Machine** | 0.9232    | 0.92240    | Piorou       |   0.93760      |

Todos melhoraram com este método, fica este

incidentes.to_csv("../input/train_data_cleaned.csv")
teste.to_csv("../input/test_data_cleaned.csv")

Todos os modelos melhoraram com este tratamento.
Neste momento já temos features com relevância para os nossos modelos, vamos, então, realizar um tratamento específico, para os top-3.
Seguir para o notebook 3.
